+++ src/documentHere/ast.ts
@@ -1,28 +1,24 @@
 import * as vs from 'vscode';
 import * as vueParser from '../utils/vue/parse';
-import { parseJS as parseTS } from '../utils/ts/babel';
-import { DocDoctor } from '../utils/ts/doc';
-import { DocDoctor4VueComponent } from '../utils/vue/doc';
+import { DocDoctor } from '../utils/ts/docDoctor';
 
-export function buildForJavaScript(editor: vs.TextEditor, forCompletion: boolean) {
-	new DocDoctor(editor).doDoc(parseTS(editor.document.getText()), forCompletion);
+export function buildForJavaScript(editor: vs.TextEditor) {
+	new DocDoctor(editor).doc();
 }
 
-export function buildForTypeScript(editor: vs.TextEditor, forCompletion: boolean) {
-	buildForJavaScript(editor, forCompletion);
+export function buildForTypeScript(editor: vs.TextEditor) {
+	buildForJavaScript(editor);
 }
 
-export function buildForVueComponent(editor: vs.TextEditor, forCompletion: boolean) {
+export function buildForVueComponent(editor: vs.TextEditor) {
 	const text = editor.document.getText();
 	// VueComponent AST
 	const vueComponentTree = vueParser.parse(text);
 	// script node
 	const scriptNode = vueParser.getContentByTagType(vueComponentTree, 'script')[0];
-	// text cleared
-	const scriptContent = text?.slice(scriptNode.contentPos, scriptNode.contentEnd);
 
-	new DocDoctor4VueComponent(editor, <number>scriptNode.contentPos, <number>scriptNode.contentEnd).doDoc(
-		parseTS(scriptContent),
-		forCompletion
-	);
+	new DocDoctor(editor).docScope({
+		start: scriptNode.contentPos as number,
+		end: scriptNode.contentEnd as number,
+	});
 }