+++ src/documentHere/index.ts
@@ -2,18 +2,18 @@ import * as vs from 'vscode';
 import * as ast from './ast';
 
 export default class DocumentHero implements vs.Disposable {
-	documentHere(editor: vs.TextEditor, commandName: string, forCompletion: boolean) {
+	documentHere(editor: vs.TextEditor, commandName: string) {
 		switch (editor.document.languageId) {
 			case 'javascript':
 			case 'javascriptreact':
-				ast.buildForJavaScript(editor, forCompletion);
+				ast.buildForJavaScript(editor);
 				break;
 			case 'typescript':
 			case 'typescriptreact':
-				ast.buildForTypeScript(editor, forCompletion);
+				ast.buildForTypeScript(editor);
 				break;
 			case 'vue':
-				ast.buildForVueComponent(editor, forCompletion);
+				ast.buildForVueComponent(editor);
 				break;
 			default:
 				vs.window.showWarningMessage(`Sorry! '${commandName}' currently do not support this file.`);