+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change Log
 
-All notable changes to the "doc-here" extension will be documented in this file.
+## [0.1.0] - 2020-11-28
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+### Added
+
+-   config for DocHere content
+-   doc for class, ts-Interface, constructor, arrowFunction
+
+### changed
+
+-   remake doc content by `typescript: ^4.1.2`
 
 ## [Unreleased]